refactor(packages): update package structure to simplify namespaces
Reorganize package structure, removing redundant "tcg.pokemon" segment for improved readability and maintainability.
